backend/requirements.txt
# =============================================================================
# GEOSPATIAL RAG - BACKEND REQUIREMENTS (Windows Compatible)
# =============================================================================
# Python 3.10+ compatible
# =============================================================================

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.30
geoalchemy2>=0.14.0

# Geospatial (use pre-built wheels)
# Install these via: pip install geopandas
# It will automatically get compatible versions of shapely, pyproj, fiona
geopandas>=0.14.0

# Note: On Windows, if NumPy fails to build, use:
# pip install --only-binary :all: numpy
# Then install other packages

# LLM Client
httpx>=0.27.0
aiohttp>=3.9.0

# Google Cloud (Voice)
google-cloud-speech>=2.26.0
google-cloud-texttospeech>=2.16.0
google-cloud-translate>=3.15.0

# Whisper (Arabic Speech-to-Text)
# Note: Requires ffmpeg installed on system
# Windows: choco install ffmpeg OR download from https://ffmpeg.org/download.html
openai-whisper>=20231117

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Audio processing
pydub>=0.25.1
numpy>=1.26.0

# Export formats
geojson>=3.1.0

# Async support
asyncio-throttle>=1.0.2

# Vector Database (RAG)
chromadb>=0.4.22

# Machine Learning (Optional - for mineral prospectivity prediction)
xgboost>=2.0.0
lightgbm>=4.0.0
category_encoders>=2.6.0
joblib>=1.3.0
scikit-learn>=1.3.0
# catboost>=1.2.0  # Optional - requires Visual Studio 2022 on Windows to build
